chore: rename Anki export files to make each one's role clear
exportAnkiDeck.js, src/apkgExport.js, and src/ankiExporter.res all
shared the same handful of words (anki/apkg/export/exporter) in
different order, making it hard to tell them apart by name alone.

- exportAnkiDeck.js -> ankiExportCli.js (the CLI entry point)
- src/apkgExport.js -> src/apkgFileExport.js (shared per-file
  orchestration: deck naming + output path, called by both the CLI
  and index.js's npm start pipeline)
- src/ankiExporter.res -> src/ankiDeckBuilder.res (the actual JSON ->
  notes/images/apkg conversion logic)

No behavior change; spec files, imports, and the export:anki script
in package.json updated to match.
