+++ package.json
@@ -16,7 +16,7 @@
   ],
   "scripts": {
     "start": "node index.js",
-    "export:anki": "node exportAnkiDeck.js",
+    "export:anki": "node ankiExportCli.js",
     "test": "NODE_OPTIONS='--experimental-vm-modules' jest",
     "test:watch": "NODE_OPTIONS='--experimental-vm-modules' jest --watch",
     "res:build": "rescript",
